lagd funksjon for å konstruere liste med alle tilgangertilskjema
